Backend/requirements.txt
# filepath: d:\projects\Queryl\Backend\requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai
tiktoken
azure-identity==1.15.0
langchain-text-splitters
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1